feat: extract protocol types into protocol.rs (#631)
Move HookInput, HookOutput, HookSpecificOutput and their serde impls
into a dedicated protocol.rs module.  main.rs drops 94 lines (3935→3841)
and the line-ratchet ceiling is lowered from 3940 to 3845.

Closes #631
