+++ .line-ratchet.toml
@@ -9,7 +9,7 @@
 
 [ratchet]
 # Current maximum allowed line count for the file.
-ceiling = 3940
+ceiling = 3845
 # The target we're ratcheting toward.
 target = 1000
 # How many lines to reduce per merge to main.
